move tx.js function to nativeTransfer.mjs and delete tx.js
